src/utils/config.py
import yaml
import json
from pathlib import Path
from typing import Dict, Any

def load_config(config_path: str) -> Dict[str, Any]:
    """Load configuration from YAML or JSON file"""
    config_path = Path(config_path)
    
    if not config_path.exists():
        raise FileNotFoundError(f"Config file not found: {config_path}")
    
    if config_path.suffix in ['.yaml', '.yml']:
        with open(config_path, 'r', encoding='utf-8') as f:
            return yaml.safe_load(f)
    elif config_path.suffix == '.json':
        with open(config_path, 'r', encoding='utf-8') as f:
            return json.load(f)
    else:
        raise ValueError(f"Unsupported config format: {config_path.suffix}")

def merge_configs(*config_paths: str) -> Dict[str, Any]:
    """Merge multiple configuration files"""
    merged_config = {}
    
    for config_path in config_paths:
        config = load_config(config_path)
        merged_config.update(config)
    
    return merged_config
