requirements.txt
# Core ML/DL dependencies
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
click>=8.0.0
python-dotenv>=1.0.0

# Experiment tracking (optional)
wandb>=0.15.0
mlflow>=2.5.0

# Development tools
jupyter>=1.0.0
ipykernel>=6.0.0
black>=23.0.0
isort>=5.12.0
pytest>=7.0.0
